refactor: remove `TextEncoderService` e reorganiza pacotes de segurança
- Remove interface `TextEncoderService` e sua implementação `TextEncoderServiceImpl`.
- Reorganiza pacotes para melhorar coesão modular, movendo elementos internos para pacotes consistentes.
- Adiciona `UserDetailsService` como um bean no `SecurityConfiguration`.

Refs: GH-1
