+++ src/main/java/dev/ghonda/project/management/security/usecases/GenerateTokenUseCase.java
@@ -1,0 +1,11 @@
+package dev.ghonda.project.management.security.usecases;
+
+import dev.ghonda.project.management.security.usecases.dto.LoginDetailPayload;
+import dev.ghonda.project.management.security.usecases.dto.LoginPayload;
+
+@FunctionalInterface
+public interface GenerateTokenUseCase {
+
+    LoginDetailPayload execute(LoginPayload payload);
+
+}
