+++ src/main/java/dev/ghonda/project/management/security/usecases/dto/LoginPayload.java
@@ -1,4 +1,4 @@
-package dev.ghonda.project.management.security.api.usecases.dto;
+package dev.ghonda.project.management.security.usecases.dto;
 
 import jakarta.validation.constraints.NotBlank;
